10th April

+++ SpringBoot/demo/src/main/java/com/basic_spring/demo/repository/PostRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.basic_spring.demo.model.Post;
 
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post,Integer> {
 }
